requirements.txt
Flask==2.3.3
xgboost==1.7.6
pandas==2.0.3
scikit-learn==1.3.0
plotly==5.15.0
numpy==1.24.3
mysql-connector-python==8.0.33
